Move FakeGTExDatabase to app; add list_genes method
