+++ app/utils/gtex_db/__init__.py
@@ -1,5 +1,6 @@
 from app.utils.gtex_db.base import GTExDatabase
 from app.utils.gtex_db.real import RealGTExDatabase
 from app.utils.gtex_db.null import NullGTExDatabase
+from app.utils.gtex_db.fake import FakeGTExDatabase
 
-__all__ = ["GTExDatabase", "RealGTExDatabase", "NullGTExDatabase"]
+__all__ = ["GTExDatabase", "RealGTExDatabase", "NullGTExDatabase", "FakeGTExDatabase"]
